Organize folder hierarchy for MVC
