+++ EssentialFeed/EssentialFeediOSTests/FeedViewControllerTests.swift
@@ -214,19 +214,19 @@ final class FeedViewControllerTests: XCTestCase {
         sut.simulateFeedImageViewNearVisible(at: 1)
         XCTAssertEqual(loader.loadedImageURLs, [image0.url, image1.url], "Expected second image URL request once second image is near visible")
     }
-    
+
     func test_feedImageView_cancelsImageURLPreloadingWhenNotNearVisibleAnymore() {
         let image0 = makeImage(url: URL(string: "http://url-0.com")!)
         let image1 = makeImage(url: URL(string: "http://url-1.com")!)
         let (loader, sut) = makeSUT()
-        
+
         sut.loadViewIfNeeded()
         loader.completeFeedLoading(with: [image0, image1])
         XCTAssertEqual(loader.cancelledImageURLs, [], "Expected no cancelled image URL requests until image is not near visible")
-        
+
         sut.simulateFeedImageViewNotNearVisible(at: 0)
         XCTAssertEqual(loader.cancelledImageURLs, [image0.url], "Expected first cancelled image URL request once first image is not near visible anymore")
-        
+
         sut.simulateFeedImageViewNotNearVisible(at: 1)
         XCTAssertEqual(loader.cancelledImageURLs, [image0.url, image1.url], "Expected second cancelled image URL request once second image is not near visible anymore")
     }
@@ -368,10 +368,10 @@ private extension FeedViewController {
         let index = IndexPath(row: row, section: feedImagesSection)
         datasource?.tableView(tableView, prefetchRowsAt: [index])
     }
-    
+
     func simulateFeedImageViewNotNearVisible(at row: Int) {
         simulateImageFeedViewVisible(at: row)
-        
+
         let datasource = tableView.prefetchDataSource
         let index = IndexPath(row: row, section: feedImagesSection)
         datasource?.tableView?(tableView, cancelPrefetchingForRowsAt: [index])
